requirements.txt
wxPython==4.2.1
pyserial==3.5
